+++ ru.taximaxim.codekeeper.ui/src/ru/taximaxim/codekeeper/ui/pgdbproject/InitProjectFromSource.java
@@ -69,6 +69,6 @@ public class InitProjectFromSource implements IRunnableWithProgress {
 
         PgDatabase db = src.get(taskpm);
         pm.newChild(25).subTask(Messages.initProjectFromSource_exporting_db_model); // 75
-        new UIProjectUpdater(db, proj).updateFull();
+        new UIProjectUpdater(db, proj).updateFull(false);
     }
 }